+++ nova-gestion-backend/src/main/java/nova/gestion/authentication/configuration/SpringSecurityConfig.java
@@ -41,9 +41,9 @@ public class SpringSecurityConfig extends WebSecurityConfigurerAdapter {
                 .csrf().disable()
                 .exceptionHandling()
                 .authenticationEntryPoint(restAuthenticationEntryPoint)
-                .and()
-                .authorizeRequests()
-                .antMatchers("/v1/**").authenticated()
+                //.and()
+                //.authorizeRequests()
+                //.antMatchers("/v1/**").authenticated()
                 .and()
                 .formLogin()
                 .successHandler(successHandler)
