+++ webui/src/app/worksheets/worksheet/worksheet.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, ViewChild} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 import {WorksheetService} from '../../services/worksheet.service';
 import {FormControl, FormGroup, Validators} from '@angular/forms';
@@ -10,12 +10,7 @@ import {StatusService} from '../../services/status.service';
 import {ProgramService} from '../../services/program.service';
 import {tap} from 'rxjs/operators';
 import {Program} from '../../models/program.model';
-import {TaskType} from '../../models/task-type';
 import {TaskService} from '../../services/task.service';
-import {countUpTimerConfigModel, CountupTimerService, timerTexts} from 'ngx-timer';
-import {Task} from 'src/app/models/task';
-import {MatDialog, MatSort, MatTableDataSource} from '@angular/material';
-import {ConfirmationDialogComponent} from '../../shared/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-worksheet',
@@ -23,42 +18,26 @@ import {ConfirmationDialogComponent} from '../../shared/confirmation-dialog/conf
   styleUrls: ['./worksheet.component.scss']
 })
 export class WorksheetComponent implements OnInit {
-  private id: any;
+  private id: number;
   worksheetForm: FormGroup;
-  taskForm: FormGroup;
   worksheet: Worksheet;
   clients: Client[] = [];
   programs: Program[] = [];
   status: Status[] = [];
-  taskTypes: TaskType[] = [];
-  timerConfig: countUpTimerConfigModel;
-  task: Task;
-  timerRunning = false;
 
-  displayedColumns = ['taskType', 'employee', 'start', 'end', 'duration', 'controls'];
-  dataSource: MatTableDataSource<Task>;
-  @ViewChild(MatSort, {static: false}) sort: MatSort;
-  selectedIndex: number;
-  startTime: any;
-  endTime: any;
-
-  constructor(public dialog: MatDialog,
-              private route: ActivatedRoute,
+  constructor(private route: ActivatedRoute,
               private worksheetService: WorksheetService,
               private router: Router,
               private clientService: ClientService,
               private programService: ProgramService,
               private taskService: TaskService,
-              private statusService: StatusService,
-              private timerService: CountupTimerService) {
+              private statusService: StatusService) {
   }
 
   ngOnInit() {
-    this.configurationTimer();
     this.initializeForm();
     this.getClients();
     this.getStatus();
-    this.getTaskTypes();
   }
 
   private initializeForm() {
@@ -71,10 +50,6 @@ export class WorksheetComponent implements OnInit {
       tool: new FormControl({value: '', disabled: true}),
       machine: new FormControl({value: '', disabled: true}),
     });
-    this.taskForm = new FormGroup({
-      taskType: new FormControl('', Validators.required),
-      time: new FormControl('')
-    });
   }
 
   private getWorksheet() {
@@ -82,14 +57,6 @@ export class WorksheetComponent implements OnInit {
       this.worksheet = res;
       this.taskService.getWorksheetTasks(this.worksheet.idWorkSheet).subscribe(tasks => {
           this.worksheet.tasks = tasks;
-          console.log(tasks);
-          this.worksheet.tasks.forEach(task => {
-            const start = new Date(task.startTime);
-            const end = new Date(task.endTime);
-            task.duration = end.getTime() - start.getTime();
-          });
-          this.dataSource = new MatTableDataSource(this.worksheet.tasks);
-          this.dataSource.sort = this.sort;
         });
       this.setValues();
     });
@@ -200,79 +167,5 @@ export class WorksheetComponent implements OnInit {
         this.worksheetForm.controls.tool.setValue(selected.tool.name);
       }
     }
-  }
-
-  private getTaskTypes() {
-    this.taskService.getAllTypes().subscribe(types => {
-      this.taskTypes = types;
-    });
-  }
-
-  start() {
-    this.timerRunning = true;
-    const date = new Date();
-    this.task = new Task();
-    this.task.startTime = date.toISOString();
-    date.setHours(date.getHours());
-    this.timerService.startTimer(date);
-  }
-
-  private stop() {
-    this.timerRunning = false;
-    this.task.endTime = new Date().toISOString();
-    this.task.idWorkSheet = this.worksheet.idWorkSheet;
-    this.timerService.stopTimer();
-    this.worksheet.tasks.push(this.task);
-  }
-
-  private configurationTimer() {
-    this.timerConfig = new countUpTimerConfigModel();
-    this.timerConfig.timerClass = 'timer';
-    this.timerConfig.timerTexts = new timerTexts();
-    this.timerConfig.timerTexts.hourText = ':';
-    this.timerConfig.timerTexts.minuteText = ':';
-    this.timerConfig.timerTexts.secondsText = ' ';
-  }
-
-  saveTask() {
-    if (this.taskForm.valid) {
-      if (this.taskForm.dirty) {
-        this.stop();
-        this.task.taskType = this.taskForm.controls.taskType.value;
-        this.taskService.add(this.task).subscribe();
-        this.getWorksheet();
-      }
-    } else {
-      this.validateAllFields(this.taskForm);
-    }
-  }
-
-  deleteTask(idTask: number) {
-    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
-      width: '350px',
-      data: 'Êtes-vous sûr de vouloir supprimer cette tâche?'
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.taskService.delete(idTask).subscribe(res => {
-          this.getWorksheet();
-        });
-      }
-    });
-  }
-
-  editTask(task: Task) {
-    this.selectedIndex = task.idTask;
-  }
-
-  updateTask(task: Task) {
-    this.selectedIndex = null;
-    task.endTime = new Date().toISOString();
-    task.startTime = new Date().toISOString();
-
-    /*this.taskService.update(task).subscribe(res => {
-      this.getWorksheet();
-    });*/
   }
 }
